Extracted filters, inclusions, exclusions, mappings and groupings to a new utils file for llmo
